requirements.txt
Django==4.2.3
djangorestframework>=3.14,<3.15
django-cors-headers>=4.0.0
Pillow>=10.0.0
django-colorfield>=0.10.1

# Production server (optional)
gunicorn>=21.2.0
whitenoise>=6.6.0

# Dev tooling (optional)
black>=24.3.0
flake8>=7.0.0
